Clean up implementation
Split extensions into multiple files to clarify what the files include.
Remove some unnecessary implementation that add complexity.
Rename files to indicate the purpose.
