+++ Sources/Shared/Extensions/Array+Extensions.swift
@@ -1,0 +1,9 @@
+public extension Array where Element : Indexable {
+
+  /// Refresh indexes inside of an array that is indexable.
+  mutating func refreshIndexes() {
+    enumerated().forEach {
+      self[$0.offset].index = $0.offset
+    }
+  }
+}
